# docs/contexts/apotek/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "BilReg API \u2014 Outpatient Apotek",
    "description": "Outpatient pharmacy API (APT-B29). Mutations require authenticated actor; role matrix remains RELEASE-BLOCKED: BC-12.",
    "version": "1.0"
  },
  "paths": {
    "/api/v1/apotek/resep-kerja/intake-electronic": {
      "post": {
        "tags": [
          "Apotek"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResepKerjaIntakeElectronicCmd"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ResepKerjaIntakeElectronicCmd"
              }
            },
            "application/*\u002Bjson": {
              "schema": {
                "$ref": "#/components/schemas/ResepKerjaIntakeElectronicCmd"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/api/v1/apotek/resep-kerja/intake-physical": {
      "post": {
        "tags": [
          "Apotek"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResepKerjaIntakePhysicalCmd"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ResepKerjaIntakePhysicalCmd"
              }
            },
            "application/*\u002Bjson": {
              "schema": {
                "$ref": "#/components/schemas/ResepKerjaIntakePhysicalCmd"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/api/v1/apotek/jual-bebas/accept": {
      "post": {
        "tags": [
          "Apotek"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JualBebasAcceptCmd"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/JualBebasAcceptCmd"
              }
            },
            "application/*\u002Bjson": {
              "schema": {
                "$ref": "#/components/schemas/JualBebasAcceptCmd"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/api/v1/apotek/jual-bebas/decline-after-accept": {
      "post": {
        "tags": [
          "Apotek"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JualBebasDeclineAfterAcceptCmd"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/JualBebasDeclineAfterAcceptCmd"
              }
            },
            "application/*\u002Bjson": {
              "schema": {
                "$ref": "#/components/schemas/JualBebasDeclineAfterAcceptCmd"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/api/v1/apotek/telaah/start": {
      "post": {
        "tags": [
          "Apotek"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TelaahStartCmd"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/TelaahStartCmd"
              }
            },
            "application/*\u002Bjson": {
              "schema": {
                "$ref": "#/components/schemas/TelaahStartCmd"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/api/v1/apotek/telaah/item": {
      "post": {
        "tags": [
          "Apotek"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TelaahUpdateItemCmd"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/TelaahUpdateItemCmd"
              }
            },
            "application/*\u002Bjson": {
              "schema": {
                "$ref": "#/components/schemas/TelaahUpdateItemCmd"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/api/v1/apotek/telaah/complete": {
      "post": {
        "tags": [
          "Apotek"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TelaahCompleteCmd"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/TelaahCompleteCmd"
              }
            },
            "application/*\u002Bjson": {
              "schema": {
                "$ref": "#/components/schemas/TelaahCompleteCmd"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/api/v1/apotek/sales-order/establish": {
      "post": {
        "tags": [
          "Apotek"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SalesOrderEstablishCmd"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/SalesOrderEstablishCmd"
              }
            },
            "application/*\u002Bjson": {
              "schema": {
                "$ref": "#/components/schemas/SalesOrderEstablishCmd"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/api/v1/apotek/sales-order/unfulfilled": {
      "post": {
        "tags": [
          "Apotek"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SalesOrderAppendUnfulfilledCmd"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/SalesOrderAppendUnfulfilledCmd"
              }
            },
            "application/*\u002Bjson": {
              "schema": {
                "$ref": "#/components/schemas/SalesOrderAppendUnfulfilledCmd"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/api/v1/apotek/sales-order/decline-purchase": {
      "post": {
        "tags": [
          "Apotek"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SalesOrderDeclinePurchaseCmd"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/SalesOrderDeclinePurchaseCmd"
              }
            },
            "application/*\u002Bjson": {
              "schema": {
                "$ref": "#/components/schemas/SalesOrderDeclinePurchaseCmd"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/api/v1/apotek/sales-order/coverage": {
      "post": {
        "tags": [
          "Apotek"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SalesOrderApplyCoverageCmd"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/SalesOrderApplyCoverageCmd"
              }
            },
            "application/*\u002Bjson": {
              "schema": {
                "$ref": "#/components/schemas/SalesOrderApplyCoverageCmd"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/api/v1/apotek/sales-order/establish-mixed": {
      "post": {
        "tags": [
          "Apotek"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SalesOrderEstablishMixedCmd"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/SalesOrderEstablishMixedCmd"
              }
            },
            "application/*\u002Bjson": {
              "schema": {
                "$ref": "#/components/schemas/SalesOrderEstablishMixedCmd"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/api/v1/apotek/sales-order/mixed-coverage": {
      "get": {
        "tags": [
          "Apotek"
        ],
        "parameters": [
          {
            "name": "resepKerjaId",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "security": []
      }
    },
    "/api/v1/apotek/queue/map": {
      "post": {
        "tags": [
          "Apotek"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QueueMapCmd"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/QueueMapCmd"
              }
            },
            "application/*\u002Bjson": {
              "schema": {
                "$ref": "#/components/schemas/QueueMapCmd"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/api/v1/apotek/queue/close": {
      "post": {
        "tags": [
          "Apotek"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QueueCloseCmd"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/QueueCloseCmd"
              }
            },
            "application/*\u002Bjson": {
              "schema": {
                "$ref": "#/components/schemas/QueueCloseCmd"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/api/v1/apotek/invoice/establish": {
      "post": {
        "tags": [
          "Apotek"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InvoiceEstablishCmd"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/InvoiceEstablishCmd"
              }
            },
            "application/*\u002Bjson": {
              "schema": {
                "$ref": "#/components/schemas/InvoiceEstablishCmd"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/api/v1/apotek/invoice/issue": {
      "post": {
        "tags": [
          "Apotek"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InvoiceIssueCmd"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/InvoiceIssueCmd"
              }
            },
            "application/*\u002Bjson": {
              "schema": {
                "$ref": "#/components/schemas/InvoiceIssueCmd"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/api/v1/apotek/invoice/payment": {
      "post": {
        "tags": [
          "Apotek"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InvoiceRecordPaymentCmd"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/InvoiceRecordPaymentCmd"
              }
            },
            "application/*\u002Bjson": {
              "schema": {
                "$ref": "#/components/schemas/InvoiceRecordPaymentCmd"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/api/v1/apotek/invoice/revise": {
      "post": {
        "tags": [
          "Apotek"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InvoiceReviseCmd"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/InvoiceReviseCmd"
              }
            },
            "application/*\u002Bjson": {
              "schema": {
                "$ref": "#/components/schemas/InvoiceReviseCmd"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/api/v1/apotek/invoice/correction": {
      "post": {
        "tags": [
          "Apotek"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InvoiceRecordCorrectionCmd"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/InvoiceRecordCorrectionCmd"
              }
            },
            "application/*\u002Bjson": {
              "schema": {
                "$ref": "#/components/schemas/InvoiceRecordCorrectionCmd"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/api/v1/apotek/invoice/correction-status": {
      "get": {
        "tags": [
          "Apotek"
        ],
        "parameters": [
          {
            "name": "invoiceId",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "security": []
      }
    },
    "/api/v1/apotek/dispensing/establish": {
      "post": {
        "tags": [
          "Apotek"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DispensingEstablishCmd"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/DispensingEstablishCmd"
              }
            },
            "application/*\u002Bjson": {
              "schema": {
                "$ref": "#/components/schemas/DispensingEstablishCmd"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/api/v1/apotek/dispensing/release": {
      "post": {
        "tags": [
          "Apotek"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DispensingReleaseCmd"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/DispensingReleaseCmd"
              }
            },
            "application/*\u002Bjson": {
              "schema": {
                "$ref": "#/components/schemas/DispensingReleaseCmd"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/api/v1/apotek/dispensing/start": {
      "post": {
        "tags": [
          "Apotek"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DispensingStartCmd"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/DispensingStartCmd"
              }
            },
            "application/*\u002Bjson": {
              "schema": {
                "$ref": "#/components/schemas/DispensingStartCmd"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/api/v1/apotek/dispensing/prepare": {
      "post": {
        "tags": [
          "Apotek"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DispensingPrepareCmd"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/DispensingPrepareCmd"
              }
            },
            "application/*\u002Bjson": {
              "schema": {
                "$ref": "#/components/schemas/DispensingPrepareCmd"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/api/v1/apotek/dispensing/pickup-call": {
      "post": {
        "tags": [
          "Apotek"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DispensingPickupCallCmd"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/DispensingPickupCallCmd"
              }
            },
            "application/*\u002Bjson": {
              "schema": {
                "$ref": "#/components/schemas/DispensingPickupCallCmd"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/api/v1/apotek/dispensing/final-review": {
      "post": {
        "tags": [
          "Apotek"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DispensingFinalReviewCmd"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/DispensingFinalReviewCmd"
              }
            },
            "application/*\u002Bjson": {
              "schema": {
                "$ref": "#/components/schemas/DispensingFinalReviewCmd"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/api/v1/apotek/dispensing/education": {
      "post": {
        "tags": [
          "Apotek"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DispensingEducationCmd"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/DispensingEducationCmd"
              }
            },
            "application/*\u002Bjson": {
              "schema": {
                "$ref": "#/components/schemas/DispensingEducationCmd"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
...
    "/api/v1/apotek/worklist/serah": {
      "get": {
        "tags": [
          "Apotek"
        ],
        "parameters": [
          {
            "name": "asOf",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "security": []
      }
    },
...
  }
}